Cancel contents animation before setting new contents in RCTNetworkImageView
Summary:
This is a fix for #322

When setting a new image via the imageURL property, the new image
doesn't always replace the previous one when it is finished downloading
because the image view has a previously instated layer animation
on its contents. This cancels any animation prior to setting the new
contents to fix the issue.
Closes https://github.com/facebook/react-native/pull/337
Github Author: Julius Parishy <julius>

Test Plan: Imported from GitHub, without a `Test Plan:` line.